% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_pipelines.R
\name{collect_learners}
\alias{collect_learners}
\title{Collect individual learners and pipelines in a \code{train_ensemble}-friendly format.}
\usage{
collect_learners(...)
}
\arguments{
\item{...}{Individual learners and pipelines to train. Lists of learners can be specified
(say outputs of \code{lrn_grid}) and are flattened into individual learners.}
}
\value{
An object of class \code{SL_Learner_Collection} to hand into \code{setup_ensemble}
}
\description{
Collect individual learners and pipelines in a \code{train_ensemble}-friendly format.
}
