% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatever_you_are.R
\name{fit_folds}
\alias{fit_folds}
\title{Fit a learner on every fold's training set}
\usage{
fit_folds(fold_list, learner, x, y, future_pkgs = character(0L))
}
\arguments{
\item{fold_list}{A \code{nana_fold_list}.}

\item{learner}{A \code{SL_Learner} object.}

\item{x}{Full predictor matrix or data frame.}

\item{y}{Full outcome vector.}

\item{future_pkgs}{Character vector of packages required by workers.}
}
\value{
A list of \code{SL_Learner_Fitted} objects, one per fold.
}
\description{
Fit a learner on every fold's training set
}
