% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{fwb_tmle_bin}
\alias{fwb_tmle_bin}
\title{Fractional Weighted Bootstrap-based TMLE, for use in nanatools}
\usage{
fwb_tmle_bin(
  treatment_model,
  or_model,
  treatment_name,
  metalearner_treatment = NULL,
  metalearner_outcome = NULL,
  trim_ipw = NULL,
  n_bstrap = 5000,
  fluctuation_family = gaussian(),
  y_bounds = NULL,
  custom_parameters = NULL,
  cluster = NULL
)
}
\arguments{
\item{treatment_model}{An object of class weightit, LazySL or glm such that applying "predict" to it provides propensity scores for a binary treatment.}

\item{or_model}{An object of class LazySL or glm. Applying "predict" should provide estimates depending on the value provided to treatment_name on the response scale.}

\item{treatment_name}{The name of the treatment variable. Has to have values in {1, 0}.}

\item{metalearner_treatment}{If the treatment model is of class LazySL, which metalearner is to be used. If NULL and the ensemble is cross-validated, defaults to the best metalearner.}

\item{metalearner_outcome}{If the outcome model is of class LazySL, which metalearner is to be used.  If NULL and the ensemble is cross-validated, defaults to the best metalearner.}

\item{trim_ipw}{Specifies how weights are to be trimmed. NULL does no trimming. "gruber" applies the data-adaptive method provided by Gruber et al. (2022).}

\item{n_bstrap}{The number of bootstrap samples used to estimate the variance of the influence curve.}

\item{fluctuation_family}{Family object used for targeting. Should match the way the outcome was fit originally.}

\item{y_bounds}{Two-length vector (a, b). If the outcome was rescaled as y_tr = (y_org - a) / (b - a), back-transforms to the original scale. If y was transformed to be \link{0,1}, this is a = min(y), b = max(y).}

\item{custom_parameters}{A list of functions, each with two arguments for the mean of treated and untreated counterfactuals, respectively. If NULL, automatically detects whether to report a log relative ATE or log odds ratio in addition to treatment-specific means and the ATE.}

\item{cluster}{Optional cluster variable by which to cluster the boostrap.}
}
\value{
A list containing a data-frame summarizing results (with at least both treatment-specific means and the ATE) and some additional objects used for fitting.
}
\description{
Fractional Weighted Bootstrap-based TMLE, for use in nanatools
}
\details{
Gruber et al. sets the lower bound of propensity scores to 5 / (sqrt(n) log(n/5)) and sets propensity scores below this threshold to this threshold.
More details in: https://doi.org/10.1093/aje/kwac087.
If weightit is used for the treatment model, its provided propensity score is used, not the weight. The provided estimand is ignored.
If LazySL is used and cv_fit_treatment is TRUE, the best ensemble is used, unless another is specified
}
\examples{
data(iris)
}
