% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_wrappers.R
\name{fwb_tmle_bin_mi}
\alias{fwb_tmle_bin_mi}
\title{A wrapper for fractional weighted bootstrap TMLE with multiply imputed data via MICE}
\usage{
fwb_tmle_bin_mi(
  treatment_models,
  or_models,
  treatment_name,
  metalearner_treatment = NULL,
  metalearner_outcome = NULL,
  trim_ipw = NULL,
  n_bstrap = 5000,
  fluctuation_family = gaussian(),
  y_bounds = NULL,
  custom_parameters = NULL,
  cluster = NULL
)
}
\arguments{
\item{treatment_name}{he name of the treatment variable. Has to have values in {1, 0}.}

\item{metalearner_treatment}{If the treatment model is of class LazySL, which metalearner is to be used. If NULL and the ensemble is cross-validated, defaults to the best metalearner.}

\item{metalearner_outcome}{Which metalearner is to be used. If NULL and the ensemble is cross-validated, defaults to the best metalearner.}

\item{trim_ipw}{Specifies how weights are to be trimmed. NULL does no trimming. "gruber" applies the data-adaptive method provided by Gruber et al. (2022).}

\item{n_bstrap}{The number of bootstrap samples used to estimate the variance of the influence curve.}

\item{fluctuation_family}{Family object used for targeting. Should match the way the outcome was fit originally.}

\item{y_bounds}{Two-length vector. If the outcome was rescaled to be in \link{0, 1}, the upper and lower bounds used. Will rescale estimands accordingly.}

\item{custom_parameters}{A list of functions, each with two arguments for the mean of treated and untreated counterfactuals, respectively. If NULL, automatically detects whether to report a log relative ATE or log odds ratio in addition to treatment-specific means and the ATE.}

\item{cluster}{An optional cluster variable to cluster the bootstrap by.}

\item{treatment_model}{An object of class LazySL_CV or weightthem. For LazySL_CV objects, applying "predict" to individual ensembles should provide estimated propensity scores.}

\item{or_model}{An object of class LazySL. Applying "predict" should provide estimates depending on the value provided to treatment_name.}
}
\value{
A list containing a data-frame summarizing results after pooling for multiple imputation, additional data frames for each imputed data set, as well as some additional objects used for fitting.
}
\description{
A wrapper for fractional weighted bootstrap TMLE with multiply imputed data via MICE
}
\examples{
data(iris)
}
