% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_1.R
\name{get_missingness_table}
\alias{get_missingness_table}
\title{Create a descriptive missingness table as data frame and tinytable-object for LaTeX exports.}
\usage{
get_missingness_table(
  data,
  subset = "missing_only",
  treatment_name = NULL,
  sig_figs = 2,
  labels = NULL,
  include_overall = "end",
  groups = NULL
)
}
\arguments{
\item{data}{A data frame containing variables that are to be summarized.}

\item{subset}{Variables for which the missingness table is to be created. Usually a character vector; default 'missing_only' uses only variables with missings, and NULL uses all variables.}

\item{treatment_name}{An optional character string containing the name of a factor variable to stratify descriptives by. Default: none.}

\item{sig_figs}{How many significant figures to print. Default: 2.}

\item{labels}{An optional character vector of labels to replace variable names in the table. Should be provided in the order provided to the order-argument.}

\item{include_overall}{If the table is stratified, should an additional non-stratified column be added? Default is 'end', making this the last column. Other options are 'beginning' and 'none'.}

\item{groups}{A named list that can optionally be specified if some variables are supposed to be grouped under a header. Each list entry contains the name of the group, and a character vector with the corresponding variables.}
}
\value{
A list containing the descriptive table as a data frame, and a tinytable-object for exporting.
}
\description{
Create a descriptive missingness table as data frame and tinytable-object for LaTeX exports.
}
\examples{
data(airquality)

get_missingness_table(
 airquality,
 subset = c("Ozone", "Solar.R", "Wind", "Temp"),
 labels = c("Ozone Level", "Solar", "Wind", "Temperature")
)



}
