% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_1.R
\name{get_table1}
\alias{get_table1}
\title{Create a descriptive table as data frame and tinytable-object for LaTeX exports.}
\usage{
get_table1(
  data,
  treatment_name = NULL,
  sig_figs = 2,
  labels = NULL,
  order = NULL,
  summarize_mean = NULL,
  summarize_median = NULL,
  summarize_binary = NULL,
  summarize_categorical = NULL,
  include_overall = "end",
  groups = NULL
)
}
\arguments{
\item{data}{A data frame containing variables that are to be summarized.}

\item{treatment_name}{An optional character string containing the name of a factor variable to stratify descriptives by. Default: none.}

\item{sig_figs}{How many significant figures to print. Default: 2.}

\item{labels}{An optional character vector of labels to replace variable names in the table. Should be provided in the order provided to the order-argument.}

\item{order}{An optional character vector of variable names with the variable ordering desired for the table. Default: column order in the provided data frame.}

\item{summarize_mean}{An optional character vector specifying which variables are to be summarized using means and standard deviations.}

\item{summarize_median}{An optional character vector specifying which variables are to be summarized using medians and interquartile ranges.}

\item{summarize_binary}{An optional character vector specifying binary factor variables to be summarized using counts and percentages.}

\item{summarize_categorical}{An optional character vector specifying factor variables to be summarized using counts and percentages.}

\item{include_overall}{If the table is stratified, should an additional non-stratified column be added? Default is 'end', making this the last column. Other options are 'beginning' and 'none'.}

\item{groups}{A named list that can optionally be specified if some variables are supposed to be grouped under a header. Each list entry contains the name of the group, and a character vector with the corresponding variables.}
}
\value{
A list containing the descriptive table as a data frame, and a tinytable-object for exporting.
}
\description{
Create a descriptive table as data frame and tinytable-object for LaTeX exports.
}
\examples{
data(iris)

get_table1(
iris,
labels = c("Sepal Length", "Sepal Width", "Petal Length", "Petal Width", "Species"),
summarize_mean = c("Sepal.Length", "Sepal.Width"),
summarize_median = c("Petal.Length", "Petal.Width"),
summarize_categorical = "Species"
)
}
