% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_constructors.R
\name{inspect.SL_Learner}
\alias{inspect.SL_Learner}
\title{Investigate an \code{SL_Learner}}
\usage{
\method{inspect}{SL_Learner}(learner)
}
\arguments{
\item{learner}{an object of class \code{SL_Learner} to inspect.}
}
\value{
Invisibly returns a list containing the original \code{fit} and \code{predict} functions.
}
\description{
Investigate an \code{SL_Learner}
}
