% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_interventions.R
\name{intervene}
\alias{intervene}
\title{Get one realization of an intervention}
\usage{
intervene(study, intervention_name)
}
\arguments{
\item{study}{A study, which is a \code{list} of class \code{nana_task}.}

\item{intervention_name}{Name of the intervention to apply to the study.}
}
\description{
Get one realization of an intervention
}
