% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearning.R
\name{lazy_cv}
\alias{lazy_cv}
\title{Title}
\usage{
lazy_cv(y, data, init)
}
\arguments{
\item{y}{Either the outcome variable itself, or a character string containing it.}

\item{data}{A data.frame or matrix containing all data (both X and y).}

\item{init}{An initialization object.}
}
\value{
A list of class lazycv containing results of the superlearner.
}
\description{
Title
}
\examples{
data(iris)
}
