% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_wrappers.R
\name{lazy_cv_mi}
\alias{lazy_cv_mi}
\title{A wrapper for lazy_cv with multiply imputed data via MICE}
\usage{
lazy_cv_mi(mi_data, y_name, var_subset = NULL, init, use_future_plan = FALSE)
}
\arguments{
\item{mi_data}{A mids-object generated via mice() containing multiply imputed data.}

\item{y_name}{The name of the outcome variable which is extracted from data sets in mi_data.}

\item{var_subset}{An optional vector of variable names which are handed into lazy_cv. Useful if not all variables used for imputations are to be used in cross-validation.}

\item{init}{The cross-validation initialization object generated via setup_learner().}

\item{use_future_plan}{If you use a future plan, set to TRUE to use it.}
}
\value{
A list of class LazyCV_SL, which contains lists of class LazyCV (one per imputed data set)
}
\description{
A wrapper for lazy_cv with multiply imputed data via MICE
}
\examples{
data(iris)
}
