% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearning.R
\name{learner_setup}
\alias{learner_setup}
\title{Title}
\usage{
learner_setup(
  learner_list,
  inner_cv = 5,
  outer_cv = 5,
  loss = "gaussian",
  metalearners = list(mtl_selector("Selected"))
)
}
\arguments{
\item{learner_list}{A list of learners, where each learner is specified as a list with entries name (character string), fit (function of data), and preds (function of fit and object)}

\item{inner_cv}{Either NULL, a positive integer indicating the amount of folds, or a function of data, for construction of the superlearner.}

\item{outer_cv}{Either NULL, a positive integer indicating the amount of folds, or a function of data, for validation of the superlearner on held-out folds.}

\item{loss}{The loss function used to estimate weights for ensembles. Either one of 'gaussian', 'binomial', and 'poisson', or a list with functions 'loss_fun' and 'loss_fun_gradient'}

\item{metalearners}{How the learners are to be ensembled. One of 'NNLogLik', 'select', or a custom function with arguments 'y', 'preds_list', and 'loss_fun_list'.}
}
\value{
An initialization object to be used for superlearning.
}
\description{
Title
}
\examples{
data(iris)
}
