% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearning_methods.R
\name{loss}
\alias{loss}
\title{Calculate losses after superlearning.}
\usage{
loss(obj, ...)
}
\arguments{
\item{obj}{An object of class lazycv.}

\item{...}{Further arguments.}
}
\description{
Calculate losses after superlearning.
}
