% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_losses.R
\name{loss_custom}
\alias{loss_custom}
\title{Create a Loss Object for Metalearning}
\usage{
loss_custom(loss, gradient = NULL, ...)
}
\arguments{
\item{loss}{A function(y, y_hat) returning a vector of losses.}

\item{gradient}{An optional function(y, y_hat) evaluating the gradient of the loss. Required for some of the pre-implemented metalearners.}

\item{...}{Additional elements (e.g., Hessian, family metadata) required by custom learners defined by users.}
}
\description{
Create a Loss Object for Metalearning
}
