% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_constructors.R
\name{lrn_custom}
\alias{lrn_custom}
\title{Create a learner factory}
\usage{
lrn_custom(fit, preds, ...)
}
\arguments{
\item{fit}{A function taking \code{x} and \code{y} that returns a fitted model.}

\item{preds}{A function taking \code{object} and \code{data} that returns predictions.}

\item{...}{Hyperparameters. Pass as \code{name = value} for defaults, or just \code{name} for required arguments.}
}
\description{
Create a learner factory
}
