% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_pipelines.R
\name{make_pipeline}
\alias{make_pipeline}
\title{Make a pipeline for use in ensembles.}
\usage{
make_pipeline(...)
}
\arguments{
\item{...}{Learners or lists of learners in the order they are to be executed.
See \code{details} to understand how to set up a valid piepline.}
}
\value{
An object of class \code{SL_Pipeline} to hand into \code{collect_learners}
}
\description{
Make a pipeline for use in ensembles.
}
\details{
\code{make_pipeline} admits both individual learners (class \code{SL_Learner}), and lists of
learners. However, it is imperative that the first argument is an \code{SL_Learner}.
The pipeline starts by training it and handing its predictions into the learner or learners
in the next argument. The output of training on a pipeline are the learner(s) specified in its last argument.
Note that, if the next argument is a list of learners, all current learners are handed into each next learner, so the output grows quickly without care.
Pipeline objects can be handed into \code{collect_learners}. Multiple pipelines can be specified,
also alongside individual \code{SL_Learner} objects.
}
