% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearning_methods.R
\name{marginals}
\alias{marginals}
\title{Caulculate marginal effects after superlearning.}
\usage{
marginals(obj, ...)
}
\arguments{
\item{obj}{An object of class lazycv.}

\item{...}{Further arguments.}
}
\description{
Caulculate marginal effects after superlearning.
}
