% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalearner_templates.R
\name{mtl_selector}
\alias{mtl_selector}
\title{A metalearner that selects the best individual learner.}
\usage{
mtl_selector(name, reduce = TRUE)
}
\arguments{
\item{name}{Name of the metalearner.}

\item{reduce}{If \code{TRUE}, collapses the prediction list into one matrix or data frame.
Useful if y is a vector and the loss function evaluates vectors. Set to FALSE for multinomial outcomes and losses.}
}
\value{
A list of class \code{SL_Metalearner_Fitted} with the fitted metalearner.
}
\description{
A metalearner that selects the best individual learner.
}
