% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_folds.R
\name{new_fold_list}
\alias{new_fold_list}
\title{Construct a list of folds}
\usage{
new_fold_list(folds)
}
\arguments{
\item{folds}{A list of \code{nana_fold} objects.}
}
\value{
An S3 object of class \code{nana_fold_list}.
}
\description{
Construct a list of folds
}
