% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearning.R
\name{setup_ensemble}
\alias{setup_ensemble}
\title{Title}
\usage{
setup_ensemble(
  learners,
  metalearners = list(mtl_selector("Selected")),
  loss = loss_gaussian()
)
}
\arguments{
\item{learners}{A list of learners, where each learner is specified as a list with entries name (character string), fit (function of data), and preds (function of fit and object)}

\item{metalearners}{How the learners are to be ensembled. Can be NULL, which makes each learner its own metalearner. \code{mtl_selector()} adds a discrete superlearner, while \code{mtl_fw()} adds a standard superlearner.}

\item{loss}{A list of class \code{mtl_loss}, containing all information the user needs for ensembling. For most cases, can use one of: \code{loss_gaussian()}, \code{loss_binomial()}, \code{loss_poisson()}, \code{loss_gamma()}}

\item{inner_cv}{Either NULL, a positive integer indicating the amount of folds, or a function of data, for construction of the superlearner.}

\item{outer_cv}{Either NULL, a positive integer indicating the amount of folds, or a function of data, for validation of the superlearner on held-out folds.}
}
\value{
An initialization object to be used for superlearning.
}
\description{
Title
}
\examples{
data(iris)
}
