% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_pipelines.R
\name{snip_paths}
\alias{snip_paths}
\title{Snip some paths of a full pipeline}
\usage{
snip_paths(object, ..., keep = FALSE)
}
\arguments{
\item{object}{Object of class \code{SL_Pipeline} within which to snip.}

\item{...}{Character vectors of length two. First entry should be the name of a
path, second entry the name of the last learner to be trained on that path.}

\item{keep}{Should the full (non-snipped) paths be retained as terminal nodes?
If you want the full non-snipped path to be trained alongside the snipped path, set to \code{TRUE}.
Default is \code{FALSE}, which discards all nodes after the snipped node.}
}
\value{
A list of class \code{SL_Pipeline}.
}
\description{
Snip some paths of a full pipeline
}
