% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_interventions.R
\name{static_intervention}
\alias{static_intervention}
\title{Static intervention: set treatment to a fixed value for all observations}
\usage{
static_intervention(value, label = NULL)
}
\arguments{
\item{value}{A single numeric value.}

\item{label}{Optional display label.}
}
\description{
Static intervention: set treatment to a fixed value for all observations
}
