% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_folds.R
\name{training_set}
\alias{training_set}
\title{Extract training indices from a fold}
\usage{
training_set(fold, ...)
}
\arguments{
\item{fold}{A \code{nana_fold}.}

\item{...}{Ignored.}
}
\description{
Extract training indices from a fold
}
