% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_folds.R
\name{validation_set}
\alias{validation_set}
\title{Extract validation indices from a fold}
\usage{
validation_set(fold, ...)
}
\arguments{
\item{fold}{A \code{nana_fold}.}

\item{...}{Ignored.}
}
\description{
Extract validation indices from a fold
}
